--- src/main/java/product/dao/ProductVO.java
@@ -1,23 +1,0 @@
-package product.dao;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.math.BigDecimal;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class ProductVO {
-    private int productId;
-    private String name;
-    private String type;
-    private int durationMonth;
-    private BigDecimal interestRate;
-
-
-}
